refactor: :truck: rename components to obey PascalCase
